Revert "Split up content header"
