--- src/content-header/ContentIconButtonGroup.react.js
@@ -1,14 +1,0 @@
-/* @flow */
-import React from 'react';
-import styles from './styles.css';
-
-export type Props = {};
-
-export default class ContentIconButtonGroup extends React.Component<Props> {
-  render() {
-    return (
-      <div className={styles.contentHeaderIconGroup}>
-        {this.props.children}
-      </div>);
-  }
-}
