--- src/content-header/styles.css
@@ -1,32 +1,0 @@
-:local(.contentHeaderRoot) {
-  border-bottom: solid #CCC 1px;
-  display: flex;
-  flex-direction: row;
-  margin: 8px 16px 0 16px;
-  padding-bottom: 8px;
-}
-
-:local(.contentHeaderIconGroup) {
-  display: flex;
-  flex-direction: row;
-  margin-right: 16px;
-}
-
-:local(.contentHeaderIconContainer) {
-  border-radius: 2px;
-  cursor: pointer;
-  margin-right: 12px;
-  padding: 4px;
-}
-
-:local(.contentHeaderIconContainer > i) {
-  color: #777;
-}
-
-:local(.contentHeaderIcon) {
-  font-size: 10pt;
-}
-
-:local(.contentHeaderIconContainer:hover) {
-  background-color: #F0F0F0;
-}
